+++ samplesheets/models.py
@@ -343,8 +343,12 @@ class Study(BaseSampleSheet):
 
     def get_nodes(self):
         """Return list of all nodes (materials and processes) for study"""
-        return list(GenericMaterial.objects.filter(study=self)) + list(
-            Process.objects.filter(study=self).prefetch_related('protocol')
+        return list(
+            GenericMaterial.objects.filter(study=self).order_by('pk')
+        ) + list(
+            Process.objects.filter(study=self)
+            .order_by('pk')
+            .prefetch_related('protocol')
         )
 
     def get_sources(self):